rag_pipeline/requirements.txt
langchain
langchain-community
datasets
sentence-transformers
langchain-chroma
langchain-huggingface
transformers
huggingface_hub
bitsandbytes
torch
hydra-core
omegaconf